+++ src/components/post/Post.tsx
@@ -28,7 +28,7 @@ export default function Post({ post }: { post: MumblePost }) {
           <h3 className={clsx('text-lg font-bold')}>{`${userData?.firstname} ${userData?.lastname}`}</h3>
           <div className={clsx('flex items-center gap-4')}>
             <Link
-              url={`/profile/${post.creator?.id}`}
+              url={`/mumble/profile/${post.creator?.id}`}
               className={'text-primary flex items-center justify-start gap-1 font-bold'}
             >
               <Profile color={'currentColor'} size={'xs'} />
